refactor: SelectedNote split into two components and refactored
