+++ src/pages/SelectedNote/SelectedNote.tsx
@@ -1,29 +1,20 @@
-import { useNavigate, useParams } from 'react-router-dom';
 import { RichTextEditor } from '@mantine/tiptap';
-import { useEditor } from '@tiptap/react';
-import { useNotes } from '@/context/NotesContext';
-import { useEffect, useRef, useState } from 'react';
-import StarterKit from '@tiptap/starter-kit';
-import Highlight from '@tiptap/extension-highlight';
-import Underline from '@tiptap/extension-underline';
-import TextAlign from '@tiptap/extension-text-align';
-import Link from '@tiptap/extension-link';
 import { TextEditorToolbar } from '@/components/TextEditor';
-import { Button, Modal, Text } from '@mantine/core';
-import { useImmediateDebouncedCallback } from '@/hooks/useImmediateDebouncedCallback';
-import { findNextArrElementIndex } from '@/helpers/findNextElementIndex';
-import { AppRoutes } from '@/types/generalTypes';
+import { Button, Text } from '@mantine/core';
 import { modals } from '@mantine/modals';
 
-export function SelectedNote() {
-  const { id } = useParams<{ id: string }>();
-  const navigate = useNavigate();
-  const { notesList, getNoteContentFromDB, updateNoteContentInDB, deleteNote } =
-    useNotes();
-  const isDeleting = useRef(false);
-  const [editable, setEditable] = useState(false);
-  const content = getNoteContentFromDB(id);
-
+interface ISelectedNote {
+  noteId: string;
+  editable: boolean;
+  setEditable: React.Dispatch<React.SetStateAction<boolean>>;
+  deleteNoteHandler: (id: string) => void;
+}
+export function SelectedNote({
+  noteId,
+  editable,
+  setEditable,
+  deleteNoteHandler,
+}: ISelectedNote) {
   const openDeleteModal = () =>
     modals.openConfirmModal({
       title: 'Delete note',
@@ -33,67 +24,8 @@ export function SelectedNote() {
       ),
       labels: { confirm: 'Delete it', cancel: "No don't delete it" },
       confirmProps: { color: 'red' },
-      onConfirm: () => deleteNoteHandler(id),
+      onConfirm: () => deleteNoteHandler(noteId),
     });
-
-  const updateHandler = (id: string, updContent: string) => {
-    debUpdateNoteContentInDB(id, updContent);
-  };
-
-  const debUpdateNoteContentInDB = useImmediateDebouncedCallback(
-    (id: string, updContent: string) => {
-      updateNoteContentInDB(id, updContent);
-    },
-    1500
-  );
-
-  const deleteNoteHandler = (id: string) => {
-    isDeleting.current = true;
-
-    const nextIndex = findNextArrElementIndex(
-      notesList,
-      notesList.findIndex(item => item.id === id)
-    );
-
-    deleteNote(id);
-
-    if (nextIndex !== -1) {
-      navigate(`/${AppRoutes.Notes}/${notesList[nextIndex].id}`);
-    } else {
-      navigate(`/${AppRoutes.Notes}`);
-    }
-
-    isDeleting.current = false;
-  };
-
-  const editor = useEditor({
-    extensions: [
-      StarterKit,
-      Underline,
-      Link,
-      Highlight,
-      TextAlign.configure({ types: ['heading', 'paragraph'] }),
-    ],
-    onUpdate: () => {
-      updateHandler(id as string, editor?.getHTML() as string);
-    },
-    content,
-  });
-
-  useEffect(() => {
-    editor?.setEditable(editable);
-  }, [editable]);
-
-  useEffect(() => {
-    // immediately saves unsaved content of prev note in DB
-    if (isDeleting.current === false) {
-      debUpdateNoteContentInDB.flush();
-    }
-    // sets content of current note 'id' to editors content
-    editor?.commands.setContent(getNoteContentFromDB(id));
-    // sets editable to false after changing note
-    setEditable(false);
-  }, [id]);
 
   return (
     <div>
@@ -106,10 +38,8 @@ export function SelectedNote() {
         Delete note
       </Button>
 
-      <RichTextEditor editor={editor}>
-        {editable && <TextEditorToolbar />}
-        <RichTextEditor.Content />
-      </RichTextEditor>
+      {editable && <TextEditorToolbar />}
+      <RichTextEditor.Content />
     </div>
   );
 }
